feat: zdf-download works with new ZDF mediathek
BREAKING CHANGE: episode data is now retrieved directly from the mediathek instead of a feed, so an ID is required instead of the feed URL.
As the URLs within the Mediathek have changed,  previous history entries are not respected so it is recommended to set a new minDate.
The repository contains an updated configuration example.